Added virtual player listener.
Added audio group listener.

Changes to support Warm Start.

Fix problem detecting whether or not position changed
when spatializing.

Throw an exception if asked for a voice bridge and none is available.

End incoming call if there are no incoming call listeners.

Use name when specified in Call instead of trying to figure out the
name from the phone number.